Add `stackit profile create` command and refactor `set` command (#351)
* new create command, refactor set command

* change flag name, improve create behaviour

* better error handling, make argument option in set

* address PR comments, improve testing for file utils copy file

* check for command existance in set and get profile

* always check if profile exists in get profile

* remove debug print

* fix formatting in error

+++ internal/pkg/config/profiles.go
@@ -11,6 +11,8 @@ import (
 	"github.com/stackitcloud/stackit-cli/internal/pkg/print"
 )
 
+const ProfileEnvVar = "STACKIT_CLI_PROFILE"
+
 // GetProfile returns the current profile to be used by the CLI.
 //
 // The profile is determined by the value of the STACKIT_CLI_PROFILE environment variable, or, if not set,
@@ -20,8 +22,8 @@ import (
 //
 // If the profile is not valid, it returns an error.
 func GetProfile() (string, error) {
-	profile, profileSet := os.LookupEnv("STACKIT_CLI_PROFILE")
-	if !profileSet {
+	profile, profileSetInEnv := GetProfileFromEnv()
+	if !profileSetInEnv {
 		contents, exists, err := fileutils.ReadFileIfExists(profileFilePath)
 		if err != nil {
 			return "", fmt.Errorf("read profile from file: %w", err)
@@ -32,11 +34,111 @@ func GetProfile() (string, error) {
 		profile = contents
 	}
 
-	err := ValidateProfile(profile)
+	// Make sure the profile exists
+	profileExists, err := ProfileExists(profile)
+	if err != nil {
+		return "", fmt.Errorf("check if profile exists: %w", err)
+	}
+	if !profileExists {
+		return "", &errors.SetInexistentProfile{Profile: profile}
+	}
+
+	err = ValidateProfile(profile)
 	if err != nil {
 		return "", fmt.Errorf("validate profile: %w", err)
 	}
 	return profile, nil
+}
+
+// GetProfileFromEnv returns the profile from the environment variable.
+// If the environment variable is not set, it returns an empty string.
+// If the profile is not valid, it returns an error.
+func GetProfileFromEnv() (string, bool) {
+	return os.LookupEnv(ProfileEnvVar)
+}
+
+// CreateProfile creates a new profile.
+// If emptyProfile is true, it creates an empty profile. Otherwise, copies the config from the current profile to the new profile.
+// If setProfile is true, it sets the new profile as the active profile.
+// If the profile already exists, it returns an error.
+func CreateProfile(p *print.Printer, profile string, setProfile, emptyProfile bool) error {
+	err := ValidateProfile(profile)
+	if err != nil {
+		return fmt.Errorf("validate profile: %w", err)
+	}
+
+	configFolderPath = filepath.Join(defaultConfigFolderPath, profileRootFolder, profile)
+
+	// Error if the profile already exists
+	_, err = os.Stat(configFolderPath)
+	if err == nil {
+		return fmt.Errorf("profile %q already exists", profile)
+	}
+
+	err = os.MkdirAll(configFolderPath, os.ModePerm)
+	if err != nil {
+		return fmt.Errorf("create config folder: %w", err)
+	}
+	p.Debug(print.DebugLevel, "created folder for the new profile: %s", configFolderPath)
+
+	currentProfile, err := GetProfile()
+	if err != nil {
+		// Cleanup created directory
+		cleanupErr := os.RemoveAll(configFolderPath)
+		if cleanupErr != nil {
+			return fmt.Errorf("get active profile: %w, cleanup directories: %w", err, cleanupErr)
+		}
+		return fmt.Errorf("get active profile: %w", err)
+	}
+
+	p.Debug(print.DebugLevel, "current active profile: %q", currentProfile)
+
+	if !emptyProfile {
+		p.Debug(print.DebugLevel, "duplicating profile configuration from %q to new profile %q", currentProfile, profile)
+		err = DuplicateProfileConfiguration(p, currentProfile, profile)
+		if err != nil {
+			// Cleanup created directory
+			cleanupErr := os.RemoveAll(configFolderPath)
+			if cleanupErr != nil {
+				return fmt.Errorf("get active profile: %w, cleanup directories: %w", err, cleanupErr)
+			}
+			return fmt.Errorf("duplicate profile configuration: %w", err)
+		}
+	}
+
+	if setProfile {
+		err = SetProfile(p, profile)
+		if err != nil {
+			return fmt.Errorf("set profile: %w", err)
+		}
+	}
+
+	return nil
+}
+
+// DuplicateProfileConfiguration duplicates the current profile configuration to a new profile.
+// It copies the config file from the current profile to the new profile.
+// If the current profile does not exist, it returns an error.
+// If the new profile already exists, it will be overwritten.
+func DuplicateProfileConfiguration(p *print.Printer, currentProfile, newProfile string) error {
+	var currentConfigFilePath string
+	// If the current profile is empty, its the default profile
+	if currentProfile == "" {
+		currentConfigFilePath = filepath.Join(defaultConfigFolderPath, fmt.Sprintf("%s.%s", configFileName, configFileExtension))
+	} else {
+		currentConfigFilePath = filepath.Join(defaultConfigFolderPath, profileRootFolder, currentProfile, fmt.Sprintf("%s.%s", configFileName, configFileExtension))
+	}
+
+	newConfigFilePath := filepath.Join(configFolderPath, fmt.Sprintf("%s.%s", configFileName, configFileExtension))
+
+	err := fileutils.CopyFile(currentConfigFilePath, newConfigFilePath)
+	if err != nil {
+		return fmt.Errorf("copy config file: %w", err)
+	}
+
+	p.Debug(print.DebugLevel, "created new configuration for profile %q based on %q in: %s", newProfile, currentProfile, newConfigFilePath)
+
+	return nil
 }
 
 // SetProfile sets the profile to be used by the CLI.
@@ -46,6 +148,15 @@ func SetProfile(p *print.Printer, profile string) error {
 		return fmt.Errorf("validate profile: %w", err)
 	}
 
+	profileExists, err := ProfileExists(profile)
+	if err != nil {
+		return fmt.Errorf("check if profile exists: %w", err)
+	}
+
+	if !profileExists {
+		return fmt.Errorf("profile %q does not exist", profile)
+	}
+
 	err = os.WriteFile(profileFilePath, []byte(profile), os.ModePerm)
 	if err != nil {
 		return fmt.Errorf("write profile to file: %w", err)
@@ -53,11 +164,6 @@ func SetProfile(p *print.Printer, profile string) error {
 	p.Debug(print.DebugLevel, "persisted new active profile in: %s", profileFilePath)
 
 	configFolderPath = filepath.Join(defaultConfigFolderPath, profile)
-	err = os.MkdirAll(configFolderPath, os.ModePerm)
-	if err != nil {
-		return fmt.Errorf("create config folder: %w", err)
-	}
-	p.Debug(print.DebugLevel, "created folder for the new profile: %s", configFolderPath)
 	p.Debug(print.DebugLevel, "profile %q is now active", profile)
 
 	return nil
@@ -88,4 +194,15 @@ func ValidateProfile(profile string) error {
 		}
 	}
 	return nil
+}
+
+func ProfileExists(profile string) (bool, error) {
+	_, err := os.Stat(filepath.Join(defaultConfigFolderPath, profileRootFolder, profile))
+	if err != nil {
+		if os.IsNotExist(err) {
+			return false, nil
+		}
+		return false, fmt.Errorf("get profile folder: %w", err)
+	}
+	return true, nil
 }
